requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
gtts==2.5.1
groq==0.11.0
httpx==0.27.2
python-dotenv==1.0.1
langdetect==1.0.9
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.6
langchain-text-splitters==0.3.0
langchain-huggingface==0.1.0
chromadb==0.5.11
sentence-transformers==3.1.1
pymupdf==1.24.10
requests==2.32.3
pydantic==2.9.2
